Prevent traffic light positions from resetting
At least on Yosemite, adding a subview to the theme frame view causes an asynchronous block to get called, that in turn calls private APIs to relayout the traffic light buttons. So, we now listen for their frame changed notifications to make sure they stay where we put them.

It's gross, but then again, this whole file is gross by necessity.